Force garbage collections after a set number of activities
Old heap where binaries are stored is not released without an explicit
garbage collection. JSP might be able to explain better/more.

We are calling GC explicitly in the variable queue, and we should do the
same here. We might not even need flow control, just regular GC so that
the process memory doesn't grow.

We are calling GC on both master & slave queues after 40 broadcast
buffer flushes

[#148892851]

Signed-off-by: Gerhard Lazu <gerhard@rabbitmq.com>